backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
iopaint==1.5.3
httpx==0.28.1
pillow==11.2.1
